rename: dummy 파일 임포트 경로 수정
