+++ src/public/views/handleCarousel.js
@@ -1,6 +1,6 @@
 import { $ } from "../utils/utils.js";
 import { CarouselContent } from "../components/CarouselContent.js";
-// import { carouselData } from "./dummy.js";
+import { carouselData } from "../../data/dummy.js";
 
 export const handleCarousel = () => {
     const carouselList = $(".carousel__list");
@@ -97,5 +97,5 @@ export const handleCarousel = () => {
 
 const renderDescription = (curIndex) => {
     const carouselContent = $('.carousel__content');
-    // carouselContent.innerHTML = CarouselContent(carouselData[curIndex - 1]);
+    carouselContent.innerHTML = CarouselContent(carouselData[curIndex - 1]);
 }
